refactoring for utils.py
